reworked cart and a lot of templates
