+++ package.json
@@ -15,7 +15,11 @@
     "del": "^0.1.3",
     "gulp-rename": "^1.2.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "gulp-declare": "^0.3.0",
+    "gulp-handlebars": "^4.0.0",
+    "gulp-wrap": "^0.11.0"
+  },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
