+++ src/js/views/cart.js
@@ -1,17 +1,14 @@
 var BasketView = Backbone.View.extend({
 	
-	template: $.Deferred(),
+	template: function(data) {
+		return Templates.cart(data);
+	},
 	
 	initialize: function() {
 		
 		var self = this;
 		
-		Einzlstck.Models.Shop.getTemplate('modules/cart.hbs').then(function(hbs) {
-			self.template.resolve(hbs);
-		});
-		
 		this.model.on('change', function() {
-			console.log('cart model changed');
 			self.render();
 		});
 		
@@ -27,8 +24,6 @@ var BasketView = Backbone.View.extend({
 			$('body').removeClass('no-scroll-mobile');
 			
 		}
-		
-		
 		
 	},
 	
@@ -54,34 +49,29 @@ var BasketView = Backbone.View.extend({
 
 		});
 		
-        // get the template
-		return this.template.then(function(hbs) {
-			
-            // render it
-			var html = hbs(data);
-			
-            // save it
-			self.el = $(html);
-			
-            // enable clicky stuff
-			self.initEvents();
-			
-			// insert into DOM
-			$('.cart').html(self.el.clone(true));
-			
-			// update totals outside of view
-			$('.total-items').text(data.total_items);
-			$('.total-price').text(data.totals.post_discount.rounded.with_tax.toFixed(2));
-			
-		});
+        // render it
+		var html = this.template(data);
+		
+		// save it
+		self.el = $(html);
+
+		// enable clicky stuff
+		self.initEvents();
+
+		// insert into DOM
+		$('.cart').html(self.el.clone(true));
+
+		// update totals outside of view
+		$('.total-items').text(data.total_items);
+		$('.total-price').text(data.totals.post_discount.rounded.with_tax.toFixed(2));
+		
+		return self.el;
 		
 	},
 	
 	initEvents: function() {
 		
 		var self = this;
-		
-		
     
 		this.el.find('.item').off('vclick');
 		this.el.find('.item').on('vclick', function(e) {
