+++ src/templates/modules/addressList.hbs
@@ -1,0 +1,11 @@
+<ul>
+    {{#each addresses}}
+    <li data-address-id="{{this.id}}">
+        <span>{{this.first_name}} {{this.last_name}}</span>
+        <span class="email">[{{this.email}}]</span><br/>
+        <span>{{this.address_1}}</span><br/>
+        <span>{{this.postcode}} {{this.city}}, {{this.country}}</span><br/>
+        <span>{{this.note}}</span>
+    </li>
+    {{/each}}
+</ul>
