+++ src/templates/modules/product.hbs
@@ -1,0 +1,17 @@
+<div class="product">
+	<div class="drop add-to-cart"><i class="fa fa-shopping-cart"></i></div>
+    <div class="details">
+		<div class="info">
+			<h3>{{title}}<span class="sold">SOLD OUT</span></h3>
+			<div class="price">{{price.value}}</div>
+		</div>
+    </div>
+    <div class="background">
+        {{#each images}}
+			{{#if @first}}
+<!--            	<div class="image active" style="background-image: url({{this.url.https}});"></div>-->
+				<img src="{{this.url.https}}">
+			{{/if}}
+        {{/each}}
+    </div>
+</div>
