+++ src/templates/pages/404.hbs
@@ -1,0 +1,33 @@
+<section class="notfound">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy="404_headline"></h1>
+                    <p data-copy="404_subheadline"></p>
+                </div>
+            </div>
+        </div>
+    </article>
+    
+    <article>
+        <div class="content">
+            <div class="units-2">
+               <div>
+                   <div class="copy center-y">
+                       <h2 data-copy="404_shopping_headline"></h2>
+                       <center><a href="/" data-copy="404_shopping_link"></a></center>
+                   </div>
+                </div>
+                <div>
+                    <div class="copy center-y">
+                        <h2 data-copy="404_insta_headline"></h2>
+                        <h3 class="social-links">
+                            <a target="_blank" href="http://www.instagram.com/einzlstck" title="Follow us on Instagram" class="insta-link"><i class="fa fa-instagram"></i></a>
+                        </h3>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+</section>
