+++ src/templates/pages/bracelets.hbs
@@ -1,0 +1,18 @@
+<section class="armbaender">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy="bracelet_headline"></h1>
+					<p>Jedes Armband ist von Hand gemacht und somit dein persönliches Unikat. Hergestellt aus solidem 925 Sterlingsilber und geschliffenen Edelsteinen, ist jedes Schmuckstück garantiert stilvoll und zugleich zeitlos. </p>
+                </div>
+            </div>
+        </div>
+    </article>
+    
+    <article>
+        <div class="content">
+            <div class="product-category" data-product-category="476"></div>
+        </div>
+    </article>
+</section>
