+++ src/templates/pages/brand.hbs
@@ -1,0 +1,114 @@
+<section class="brand">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1>Wir sind Einzelstück</h1>
+                    <h3>Lerne uns kennen!</h3>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        
+						<h2>Echt. Handgemacht. Einzigartig.</h2>
+						<p>
+							Unsere Schmuckstücke bestehen aus 925 Sterlingsilber und echten Edel- und Halbedelsteinen.
+							Von Hand geformte Fassungen verzieren wunderschöne, handgeschliffene Edelsteine.<br/>
+							Wie wir, besitzt jedes Schmuckstück eine Geschichte die es geprägt hat. Handarbeitsbedingte Besonderheiten machen jedes Stück zu einem Einzelstück.
+						</p>
+                            
+                        
+                    </div>
+                </div>
+                <div>
+                    <div class="background">
+						<div class="image active" style="background-image: url(images/backgrounds/handmade.jpg)"></div>
+					</div>
+                </div>
+            </div>
+        </div>
+    </article>
+	<article>
+        <div class="content">
+            <div class="units-2">
+				<div>
+                    <div class="background">
+						<div class="image active" style="background-image: url(images/backgrounds/tradition.jpg)"></div>
+					</div>
+                </div>
+                <div>
+                    <div class="copy">
+                        
+						<h2>Tradition aus Zeiten der Maharadschas</h2>
+						<p>
+							Unser Schmuck wird von Familien mit langer Tradition in der Edelmetal- und Edelsteinverarbeitung gefertigt. 
+							Bis 1947 fertigten sie Schmuck für die Könige Indiens, sogenannte Maharadschas. 
+							Jede Familie kümmerte sich um die Verarbeitung eines bestimmten Edelsteins.
+							Als 1947 die Herrschaft der Maharadschas endete, fertigten diese Familien Schmuck für westliche Großhändler und Touristen.<br/>
+							Wir lassen unseren Schmuck jedoch direkt bei diesen Familien herstellen. Es gibt keine Großhändler dazwischen womit wir 
+							überdurchschnittliche hohe und faire Preise zahlen können, um den indischen Familien die Unabhängikeit von Großhändlern zu erleichtern.
+						</p>
+                            
+                        
+                    </div>
+                </div>
+                
+            </div>
+        </div>
+    </article>
+	<article>
+        <div class="content">
+            <div class="units-2">
+				
+                <div>
+                    <div class="copy">
+                        
+						<h2>Die Vision</h2>
+						<p>
+							Jaipur, Rajasthan, Indien ist die Heimat der traditionsreichen Familien, die unseren Schmuck herstellen.<br/>
+							Um diese Familien vor den Niedrigstpreisen der Großhändler zu schützen, handeln wir mit ihnen ohne Umwege und auf einer fairen Basis, doch wir möchten mehr.<br/>
+							Unser Traum ist es, eines Tages die Bildung der Kinder dieser Familien finanzieren zu können.
+						</p>
+                            
+                        
+                    </div>
+                </div>
+                <div>
+                    <div class="background">
+						<div class="image active" style="background-image: url(images/backgrounds/vision.jpg)"></div>
+					</div>
+                </div>
+            </div>
+        </div>
+    </article>
+	<article>
+        <div class="content">
+            <div class="units-2">
+				<div>
+                    <div class="background">
+						<div class="image active" style="background-image: url(images/backgrounds/wir.jpg); background-size: contain;"></div>
+					</div>
+                </div>
+                <div>
+                    <div class="copy">
+                        
+						<h2>Wir</h2>
+						<p>
+							Lange war Einzelstück eine Idee und Zukunftsvision, inspiriert von ähnlichen Konzepten und FairTrade Bestrebungen in New York. 
+							2014 begannen Marina und Sumit mit der Umsetzung dieser Idee im Stuttgarter Westen.
+						</p>
+                        
+                        
+                    </div>
+                </div>
+                
+            </div>
+        </div>
+    </article>
+	<center style="padding: 20px;"><i class="fa fa-heart"></i></center>
+</section>
