+++ src/templates/pages/checkout.hbs
@@ -1,0 +1,58 @@
+<section class="checkout">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy="checkout_headline"></h1>
+                    <p data-copy="checkout_subheadline"></p>
+                </div>
+            </div>
+        </div>
+    </article>
+    
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div id="billing-address" class="unit">
+                    <div class="copy">
+                        <h2 data-copy="checkout_billing_headline"></h2>
+                        <div class="user-address-list"></div>
+                        <button class="new-address-button">
+                            <span data-copy="checkout_add_address_button"></span>
+                        </button>
+						<div class="new-address-form"></div>
+                        
+                    </div>
+                </div>
+                <div id="shipping-address" class="unit">
+                    <div class="copy">
+                        <h2 data-copy="checkout_shipping_headline"></h2>
+                        <div class="user-address-list"></div>
+						<button class="new-address-button">
+                            <span data-copy="checkout_add_address_button"></span>
+                        </button>
+                        <div class="new-address-form"></div>
+                    </div>
+                </div>
+            </div>
+            <div class="units-2">
+                <div id="cart-preview">
+                    <div class="copy">
+                        <h2 data-copy="checkout_cart_headline"></h2>
+                        <div class="cart"></div>
+                    </div>
+                </div>
+                <div id="payment-choice">
+                    <div class="copy">
+                        <h2 data-copy="checkout_pay_headline"></h2>
+                        <ul>
+                            <li><input type="radio" name="paymentOption" id="paypal" value="paypal-express" checked /><label for="paypal" data-copy="checkout_pay_option1"></label></li>
+                            <li><input type="radio" name="paymentOption" id="transfer" value="manual" /><label for="transfer">Vorkasse</label></li>
+                        </ul>
+                        <button class="button buy"><span data-copy="checkout_process_order_button"></span></button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+</section>
