+++ src/templates/pages/f12.hbs
@@ -1,0 +1,34 @@
+<section class="f12">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1>// just a note</h1>
+                    <div class="limitWidth">
+                        <p></p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        <h2>This is for the devs</h2>
+                        <p>If you're a developer, an ecommerce business owner or just curious - you found the right place.</p>
+                        <p>This page tells you a bit more about the technology, techniques and systems we use on Einzelstück.</p>
+                    </div>
+                </div>
+                <div>
+                    <div class="copy">
+                        <h2>Meep</h2>
+                        <p>Möp Möp</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+    </article>
+</section>
