+++ src/templates/pages/home.hbs
@@ -1,0 +1,110 @@
+<section class="home">
+    <article>
+        <div class="units-1">
+            <div>
+                <div class="copy center-y">
+                    <h1 class="stretch" data-copy="title_long"></h1>
+                    <div class="limit-width">
+                        <p data-copy="site_desc"></p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+				<div>
+                    <img style="width: 100%;" src="images/backgrounds/vision.jpg">
+                </div>
+				<div>
+					<div class="link">
+						<a href="/brand">
+							<div class="link-text">Lerne uns kennen! &rarr;</div>
+						</a>
+					</div>
+				</div>
+            </div>
+            <div class="units-2">
+				<div>
+					<div class="link">
+						<a href="/rings">
+							<div class="link-text" data-copy="link_rings"></div>
+						</a>
+					</div>
+				</div>
+				<div data-product="11815"></div>
+            </div>
+        </div>
+    </article>
+<!--
+	<article>
+		<div class="content">
+			<div class="units-1">
+				<div>
+					<div class="background">
+						<div class="image active" style="background-image: url(images/backgrounds/sunset-hair.jpg)"></div>
+					</div>
+				</div>
+			</div>
+		</div>
+        
+    </article>
+-->
+	<article>
+        <div class="content">
+            <div class="units-2">
+				<div data-product="973348699570700709"></div>
+                <div>
+					<div class="link">
+						<a href="/bracelets">
+							<div class="link-text" data-copy="link_bracelets"></div>
+						</a>
+					</div>
+				</div>
+                
+				<div>
+					<div class="link">
+						<a href="/pendants">
+							<div class="link-text">Alle Anhänger &rarr;</div>
+						</a>
+					</div>
+				</div>
+				<div data-product="921699135906644033"></div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        <form class="newsletter-form">
+                        <h2>Newsletter</h2>
+						<h3>Trage dich ein um Neuigkeiten, fantastische Angebote und Infos zu Rabatt-Aktionen zu erhalten.</h3>
+						<input type="email" placeholder="E-Mail eintragen">
+						<button type="submit" class="subscribe">
+							<span data-copy="nl_button"></span>
+						</button>
+                    </form>
+                    </div>
+                </div>
+                <div>
+                    <div class="copy">
+                        <h2 data-copy="follow_headline"></h2>
+                        <p data-copy="follow_copy1"></p>
+                        <p data-copy="follow_copy2"></p>
+                        <br><br>
+                        <h3 class="social-links">
+							<a target="_blank" href="http://www.facebook.com/einzlstck" title="Like us on Facebook" class="facebook-link"><i class="fa fa-facebook"></i></a>
+<!--                            <a target="_blank" href="http://www.twitter.com/einzlstck" title="Follow us on Twitter" class="twitter-link"><i class="fa fa-twitter"></i></a>-->
+                            <a target="_blank" href="http://www.instagram.com/einzlstck" title="Follow us on Instagram" class="insta-link"><i class="fa fa-instagram"></i></a>
+                        </h3>
+                        
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+    
+</section>
