+++ src/templates/pages/imprint.hbs
@@ -1,0 +1,98 @@
+<section class="imprint">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1>Impressum</h1>
+                    <h3>und Haftungsausschluss</h3>
+                </div>
+            </div>
+        </div>
+    </article>
+    <center style="padding: 20px;"><small>Angaben gemäß § 5 TMG</small></center>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        
+						<h2>Inhaltlich Verantwortlicher</h2>
+						<p>
+							Sumit Kumar<br/>
+							Schwabstr. 32<br/>
+							70197 Stuttgart
+						</p>
+						<p>
+							Telefon: +4917664104083<br/>
+							E-Mail: info@einzlstck.de
+						</p>
+                        <p>
+							Gemäß §19 des Umsatzsteuergesetzes verzichten wir nach Kleinunternehmerregelung auf Erhebung und Ausweisung der Umsatzsteuer.
+						</p>
+                        
+                    </div>
+                </div>
+                <div>
+                    
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        <h2>Haftung für Inhalte</h2>
+                        <p>Als Diensteanbieter sind wir gemäß § 7 Abs.1 TMG für eigene Inhalte auf diesen Seiten nach den allgemeinen Gesetzen verantwortlich. Nach §§ 8 bis 10 TMG sind wir als Diensteanbieter jedoch nicht verpflichtet, übermittelte oder gespeicherte fremde Informationen zu überwachen oder nach Umständen zu forschen, die auf eine rechtswidrige Tätigkeit hinweisen. Verpflichtungen zur Entfernung oder Sperrung der Nutzung von Informationen nach den allgemeinen Gesetzen bleiben hiervon unberührt. Eine diesbezügliche Haftung ist jedoch erst ab dem Zeitpunkt der Kenntnis einer konkreten Rechtsverletzung möglich. Bei Bekanntwerden von entsprechenden Rechtsverletzungen werden wir diese Inhalte umgehend entfernen.</p>
+                    </div>
+                </div>
+                <div>
+                    <div class="copy center-y">
+                        <h2>Wir sind Bäcker</h2>
+                        <h3>Leider wissen wir nicht immer, wo die Butter her kommt.</h3>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        <h2>Haftung für Links</h2>
+                        <p>Unser Angebot enthält Links zu externen Webseiten Dritter, auf deren Inhalte wir keinen Einfluss haben. Deshalb können wir für diese fremden Inhalte auch keine Gewähr übernehmen. Für die Inhalte der verlinkten Seiten ist stets der jeweilige Anbieter oder Betreiber der Seiten verantwortlich. Die verlinkten Seiten wurden zum Zeitpunkt der Verlinkung auf mögliche Rechtsverstöße überprüft. Rechtswidrige Inhalte waren zum Zeitpunkt der Verlinkung nicht erkennbar. Eine permanente inhaltliche Kontrolle der verlinkten Seiten ist jedoch ohne konkrete Anhaltspunkte einer Rechtsverletzung nicht zumutbar. Bei Bekanntwerden von Rechtsverletzungen werden wir derartige Links umgehend entfernen.</p>
+                    </div>
+                </div>
+                <div>
+                    <div class="copy center-y">
+                        <h2>Wir sind Lokführer</h2>
+                        <h3>Wir bieten dir die Fahrt an, nicht das Ziel.</h3>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+                <div>
+                    <div class="copy">
+                        <h2>Urheberrecht</h2>
+                        <p><del>Die durch die Seitenbetreiber erstellten Inhalte und Werke auf diesen Seiten unterliegen dem deutschen Urheberrecht. Die Vervielfältigung, Bearbeitung, Verbreitung und jede Art der Verwertung außerhalb der Grenzen des Urheberrechtes bedürfen der schriftlichen Zustimmung des jeweiligen Autors bzw. Erstellers. Downloads und Kopien dieser Seite sind nur für den privaten, nicht kommerziellen Gebrauch gestattet.</del></p>
+                        <p>Soweit die Inhalte auf dieser Seite nicht vom Betreiber erstellt wurden, werden die Urheberrechte Dritter beachtet. Insbesondere werden Inhalte Dritter als solche gekennzeichnet. Sollten Sie trotzdem auf eine Urheberrechtsverletzung aufmerksam werden, bitten wir um einen entsprechenden Hinweis. Bei Bekanntwerden von Rechtsverletzungen werden wir derartige Inhalte umgehend entfernen.</p>
+                    </div>
+                </div>
+                <div>
+                    <div class="copy center-y">
+                        <h2 class="quatsch-arrow">Quatsch mit Soße</h2>
+                        <h3>Tu es, kopiere uns</h3>
+                        <p>Hiermit erlauben wir dir unsere Bilder unter der <a href="https://creativecommons.org/licenses/by/3.0/de/" target="_blank">Creative Commons Lizens</a> zu benutzen und unsere Texte zu kopieren. Sogar der komplette Code unserer Website ist <a href="https://github.com/codeofsumit/einzl" target="_blank" title="Einzelstück on Github">Open Source <i class="fa fa-github"></i></a> unter der <a href="http://opensource.org/licenses/MIT" title="MIT License" target="_blank">MIT Licence</a>.<br/>Benutze was du möchtest, wofür du möchtest.</p>
+                        <p>Außerdem klauen wir nicht von anderen. Sollten wir es unbewusst getan haben, lass es uns wissen!</p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+</section>
