+++ src/templates/pages/likes.hbs
@@ -1,0 +1,30 @@
+<section class="likes">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy=""><i class="fa fa-heart"></i></h1>
+                    <p>The Ones That Conquered Your Heart</p>
+                </div>
+            </div>
+        </div>
+    </article>
+<!--
+    <article>
+        <div class="content">
+            <div class="units-4">
+                <div>
+                    <div class="copy">
+                        <h2 data-copy="furnish_headline"></h2>
+                        <p data-copy="furnish_copy"></p>
+                    </div>
+                </div>
+                <div data-product="11816"></div>
+                <div data-product="11806"></div>
+                <div data-product="11810"></div>
+            </div>
+        </div>
+    </article>
+-->
+    
+</section>
