+++ src/templates/pages/necklaces.hbs
@@ -1,0 +1,18 @@
+<section class="ketten">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy="necklace_headline"></h1>
+					<p data-copy="like_copy"></p>
+                </div>
+            </div>
+        </div>
+    </article>
+    
+    <article>
+        <div class="content">
+            <div class="product-category" data-product-category="478"></div>
+        </div>
+    </article>
+</section>
