+++ src/templates/pages/pendants.hbs
@@ -1,0 +1,20 @@
+<section class="pendants">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy="pendants_headline"></h1>
+					<p>
+						Wie aufblühende Blumen umfassen die 925 Sterlingsilber-Fassungen dieser Anhänger ihre geschliffenen, echten Edelsteine.
+						Von Hand verarbeitet wird jeder Anhänger zu einem Einzelstück.
+					</p>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="product-category" data-product-category="949"></div>
+        </div>
+    </article>
+</section>
