+++ src/templates/pages/rings.hbs
@@ -1,0 +1,17 @@
+<section class="ringe">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1 data-copy="rings_headline"></h1>
+					<p>All unsere Ringe bestehen aus kunstvoll verzierten 925 Sterlingsilber-Fassungen die echte geschliffene Edelsteine umfassen.</p>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="product-category" data-product-category="477"></div>
+        </div>
+    </article>
+</section>
