+++ src/templates/pages/shipping.hbs
@@ -1,0 +1,119 @@
+<section class="shipping">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1>Versand & Rückgabe</h1>
+                </div>
+            </div>
+        </div>
+    </article>
+<!--    <center style="padding: 20px;"><small>Angaben gemäß § 5 TMG</small></center>-->
+    <article>
+        <div class="content">
+            <div class="units-2">
+				<div>
+					
+					<div class="copy">
+						<h2>Versand</h2>
+						
+						<p>
+							Aktuell versenden wir ausschließlich innerhalb Deutschlands, nach Österreich und in die Schweiz.
+							Wir beauftragen DHL mit der Lieferung.
+						</p>
+					</div>
+					
+				</div>
+                <div>
+                    <div class="copy">
+						
+						<h2>Widerrufsbelehrung</h2>
+						
+						<p>
+							Sie haben das Recht, binnen vierzehn Tagen ohne Angabe von Gründen diesen Vertrag zu widerrufen. 
+							Die Widerrufsfrist beträgt vierzehn Tage ab dem Tag, an dem Sie oder ein von Ihnen benannter Dritter, der nicht der Beförderer ist, 
+							die Waren in Besitz genommen haben bzw. hat. Um Ihr Widerrufsrecht auszuüben, 
+							müssen Sie uns, Einzelstück - Sumit Kumar, Schwabstraße 32, 70197 Stuttgart, E-Mail: info@einzlstck.de 
+							mittels einer eindeutigen Erklärung (z. B. ein mit der Post versandter Brief, Telefax oder E-Mail) über Ihren Entschluss, 
+							diesen Vertrag zu widerrufen, informieren.
+						</p>
+						
+						<p>
+							Sie können dafür das beigefügte Muster-Widerrufsformular verwenden, dessen Verwendung jedoch nicht vorgeschrieben ist. Zur Wahrung der Widerrufsfrist reicht es aus, dass Sie die Mitteilung über die Ausübung des Widerrufsrechts vor Ablauf der Widerrufsfrist absenden.
+						</p>
+					</div>
+				</div>
+				
+				
+			</div>
+			<div class="units-2">
+				<div>
+					<div class="copy">
+						<h2>Versandkosten</h2>
+						<p>
+							Der Käufer trägt keine Versandkosten. Die Versandkosten aller Lieferungen in unsere aktuellen Liefergebiete werden von uns übernommen.
+						</p>
+					</div>
+				</div>
+				<div>
+                    <div class="copy">
+						<h2>Widerrufsfolgen</h2>
+						
+						<p>
+							Wenn Sie diesen Vertrag widerrufen, haben wir Ihnen alle Zahlungen, die wir von Ihnen erhalten haben, einschließlich der Lieferkosten (mit Ausnahme der zusätzlichen Kosten, die sich daraus ergeben, dass Sie eine andere Art der Lieferung als die von uns angebotene, günstigste Standardlieferung gewählt haben), unverzüglich und spätestens binnen vierzehn Tagen ab dem Tag zurückzuzahlen, an dem die Mitteilung über Ihren Widerruf dieses Vertrags bei uns eingegangen ist.
+						</p>
+						<p>
+							Für diese Rückzahlung verwenden wir dasselbe Zahlungsmittel, das Sie bei der ursprünglichen Transaktion eingesetzt haben, es sei denn, mit Ihnen wurde ausdrücklich etwas anderes vereinbart; in keinem Fall werden Ihnen wegen dieser Rückzahlung Entgelte berechnet.
+						</p>
+						<p>
+							Wir können die Rückzahlung verweigern, bis wir die Waren wieder zurückerhalten haben oder bis Sie den Nachweis erbracht haben, dass Sie die Waren zurückgesandt haben, je nachdem, welches der frühere Zeitpunkt ist.
+						</p>
+						<p>
+							Sie haben die Waren unverzüglich und in jedem Fall spätestens binnen vierzehn Tagen ab dem Tag, an dem Sie uns über den Widerruf dieses Vertrags unterrichten, an uns zurückzusenden oder zu übergeben. Die Frist ist gewahrt, wenn Sie die Waren vor Ablauf der Frist von vierzehn Tagen absenden.
+						</p>
+						<p>
+							Sie tragen die unmittelbaren Kosten der Rücksendung der Waren. Sie müssen für einen etwaigen Wertverlust der Waren nur aufkommen, wenn dieser Wertverlust auf einen zur Prüfung der Beschaffenheit, Eigenschaften und Funktionsweise der Waren nicht notwendigen Umgang durch Sie zurückzuführen ist.
+						</p>
+						
+					</div>
+				</div>
+				
+			</div>
+			<div class="units-2">
+                <div>
+					<div class="copy">
+						<h2>Lieferzeit</h2>
+						<p>
+							Aufgrund unserer individueller Pflege und Verpackung für jeden einzelnen Kunden kann unsere Lieferzeit bis zu drei Wochen in Anspruch nehmen.
+							In der Regel sollte die Bestellung jedoch nach einer Woche eintreffen.
+						</p>
+					</div>
+				</div>
+				<div>
+                    <div class="copy">
+						<h2>Muster-Widerrufsformular</h2>
+						
+						<p>(Wenn Sie den Vertrag widerrufen wollen, dann füllen Sie bitte dieses Formular aus und senden Sie es zurück.)</p>
+						
+						<blockquote>
+							
+							<p>An<br/>Einzelstück - Sumit Kumar<br/>Schwabstraße 32<br/>70197 Stuttgart,<br/>E-Mail: info@einzlstck.de</p>
+							
+							<p>Hiermit widerrufe(n) ich/wir (*) den von mir/uns (*) abgeschlossenen Vertrag über den Kauf der folgenden Waren (*)/die Erbringung der folgenden Dienstleistung (*)</p>
+							<p>Bestellt am (*)/erhalten am (*)</p>
+							<p>Name des/der Verbraucher(s)</p>
+							<p>Anschrift des/der Verbraucher(s)</p>
+							<p>Unterschrift des/der Verbraucher(s) (nur bei Mitteilung auf Papier)</p>
+							<p>Datum</p>
+							<p>(*) Unzutreffendes streichen.</p>
+						</blockquote>
+
+
+
+
+                    </div>
+                </div>
+            </div>
+        </div>
+    </article>
+</section>
