+++ src/templates/pages/testpage.hbs
@@ -1,0 +1,25 @@
+<section class="testpage">
+    <article>
+        <div class="units-1 cut-in-half">
+            <div>
+                <div class="copy center-y">
+                    <h1>Test Page</h1>
+                </div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-2">
+				<div data-product="992712251494367443"></div>
+                <div></div>
+            </div>
+        </div>
+    </article>
+    <article>
+        <div class="content">
+            <div class="units-4">
+            </div>
+        </div>
+    </article>
+</section>
